+++ test/providers/presenter_widget_test.dart
@@ -27,13 +27,47 @@ void main() {
 
     expect(presenter.outputUpdateLogs, ['a', 'b']);
   });
+
+  testWidgets(
+    'didUpdatePresenter test',
+    (tester) async {
+      final widget = MaterialApp(
+        home: FutureBuilder<int>(
+          initialData: 1,
+          future: Future.delayed(const Duration(milliseconds: 100), () => 2),
+          builder: (context, snapshot) {
+            return TestPresenter(
+              count: snapshot.data,
+              builder: (viewModel) => Text(viewModel.foo, key: Key('foo')),
+            );
+          },
+        ),
+      );
+
+      await ProviderTester().pumpWidget(tester, widget);
+
+      final testPresenterFinder = find.byType(TestPresenter);
+      expect(testPresenterFinder, findsOneWidget);
+
+      var presenter = tester.widget<TestPresenter>(testPresenterFinder);
+      expect(presenter.didUpdatePresenterLogs, isEmpty);
+
+      await tester.pump(const Duration(milliseconds: 100));
+      presenter = tester.widget<TestPresenter>(testPresenterFinder);
+      expect(presenter.didUpdatePresenterLogs.first.previous, 1);
+      expect(presenter.didUpdatePresenterLogs.first.next, 2);
+    },
+  );
 }
 
 class TestPresenter extends Presenter<TestViewModel, TestOutput, TestUseCase> {
-  TestPresenter({required PresenterBuilder<TestViewModel> builder})
+  TestPresenter({required PresenterBuilder<TestViewModel> builder, this.count})
       : super(provider: provider, builder: builder);
 
+  final int? count;
+
   final List<String> outputUpdateLogs = [];
+  final List<PresenterData> didUpdatePresenterLogs = [];
 
   @override
   void onLayoutReady(BuildContext context, TestUseCase useCase) {
@@ -48,6 +82,16 @@ class TestPresenter extends Presenter<TestViewModel, TestOutput, TestUseCase> {
   void onOutputUpdate(BuildContext context, TestOutput output) {
     super.onOutputUpdate(context, output);
     outputUpdateLogs.add(output.foo);
+  }
+
+  @override
+  void didUpdatePresenter(
+    BuildContext context,
+    TestPresenter old,
+    TestUseCase useCase,
+  ) {
+    super.didUpdatePresenter(context, old, useCase);
+    didUpdatePresenterLogs.add(PresenterData(old.count, count));
   }
 }
 
@@ -87,4 +131,11 @@ class TestViewModel extends ViewModel {
 
   @override
   List<Object?> get props => [foo];
+}
+
+class PresenterData {
+  PresenterData(this.previous, this.next);
+
+  final int? previous;
+  final int? next;
 }